feature: ignore unsupported YAML files (#6)
This adds logic to ignore unsupported YAML files (non-GitHub Workflows, and non-GitHub Composite Actions YAML files). This new logic is also enabled by default.

It makes it easier to work with `shellcheck-gha` as `.github/` directory may contains multiple kinds of YAML files such as `dependabot.yaml`. Before this change it would require to make sure there are no unrelated YAML files in the `.github/` directory which is inconvenient and makes the tool hard to use.